fix: 상품 엔티티와 옵션 엔티티 단방향 관계로 변경

+++ src/main/java/gift/service/OptionService.java
@@ -1,6 +1,8 @@
 package gift.service;
 
 import static gift.util.constants.OptionConstants.OPTION_NOT_FOUND;
+import static gift.util.constants.OptionConstants.OPTION_NAME_DUPLICATE;
+import static gift.util.constants.OptionConstants.OPTION_REQUIRED;
 import static gift.util.constants.ProductConstants.PRODUCT_NOT_FOUND;
 
 import gift.dto.option.OptionCreateRequest;
@@ -27,10 +29,8 @@ public class OptionService {
     }
 
     public List<OptionResponse> getOptionsByProductId(Long productId) {
-        Product product = productRepository.findById(productId)
-            .orElseThrow(() -> new ProductNotFoundException(PRODUCT_NOT_FOUND + productId));
-
-        return product.getOptions().stream()
+        List<Option> options = optionRepository.findByProductId(productId);
+        return options.stream()
             .map(OptionService::convertToDTO)
             .toList();
     }
@@ -51,6 +51,8 @@ public class OptionService {
         Product product = productRepository.findById(productId)
             .orElseThrow(() -> new ProductNotFoundException(PRODUCT_NOT_FOUND + productId));
 
+        validateDuplicateOptionName(productId, optionCreateRequest.name());
+
         Option option = new Option(null, optionCreateRequest.name(), optionCreateRequest.quantity(),
             product);
         Option savedOption = optionRepository.save(option);
@@ -69,6 +71,8 @@ public class OptionService {
             throw new OptionNotFoundException(OPTION_NOT_FOUND + optionId);
         }
 
+        validateDuplicateOptionName(productId, optionUpdateRequest.name());
+
         option.update(optionUpdateRequest.name(), optionUpdateRequest.quantity(), product);
         Option updatedOption = optionRepository.save(option);
         return convertToDTO(updatedOption);
@@ -84,7 +88,20 @@ public class OptionService {
             throw new OptionNotFoundException(OPTION_NOT_FOUND + optionId);
         }
 
+        if (optionRepository.findByProductId(productId).size() == 1) {
+            throw new IllegalArgumentException(OPTION_REQUIRED);
+        }
+
         optionRepository.delete(option);
+    }
+
+    private void validateDuplicateOptionName(Long productId, String optionName) {
+        List<Option> options = optionRepository.findByProductId(productId);
+        for (Option option : options) {
+            if (option.isNameMatching(optionName)) {
+                throw new IllegalArgumentException(OPTION_NAME_DUPLICATE);
+            }
+        }
     }
 
     // Mapper methods
@@ -92,7 +109,8 @@ public class OptionService {
         return new OptionResponse(
             option.getId(),
             option.getName(),
-            option.getQuantity()
+            option.getQuantity(),
+            option.getProduct().getId()
         );
     }
 }
